+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Save a backup of the database.
 - Load a backup of the database.
 
-## [2.0.0] - 2023-02-27 (expected)
+## [2.0.0] - 2023-02-26
 
 ### Added
 
@@ -71,4 +71,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Project README.
 - .gitignore
 
+[2.0.0]: https://github.com/4N0NYM0U5MY7H/CS361_Individual_Project/releases/tag/v2.0.0
 [1.0.0]: https://github.com/4N0NYM0U5MY7H/CS361_Individual_Project/releases/tag/v1.0.0